+++ packages/react-dom/src/client/ReactDOMComponent.js
@@ -9,7 +9,6 @@
 
 import {registrationNameModules} from 'legacy-events/EventPluginRegistry';
 import {canUseDOM} from 'shared/ExecutionEnvironment';
-import endsWith from 'shared/endsWith';
 import invariant from 'shared/invariant';
 import {setListenToResponderEventTypes} from '../events/DeprecatedDOMEventResponderSystem';
 
@@ -1325,6 +1324,11 @@ export function restoreControlledState(
       ReactDOMSelectRestoreControlledState(domElement, props);
       return;
   }
+}
+
+function endsWith(subject: string, search: string): boolean {
+  const length = subject.length;
+  return subject.substring(length - search.length, length) === search;
 }
 
 export function listenToEventResponderEventTypes(
