+++ packages/react-dom/src/events/SelectEventPlugin.js
@@ -7,7 +7,7 @@
 
 import {canUseDOM} from 'shared/ExecutionEnvironment';
 import SyntheticEvent from 'legacy-events/SyntheticEvent';
-import isTextInputElement from 'shared/isTextInputElement';
+import isTextInputElement from './isTextInputElement';
 import shallowEqual from 'shared/shallowEqual';
 
 import {
