+++ packages/react-interactions/events/src/dom/Press.js
@@ -7,7 +7,7 @@
  * @flow
  */
 
-import type {PointerType} from 'shared/ReactDOMTypes';
+import type {PointerType} from 'react-dom/src/shared/ReactDOMTypes';
 
 import * as React from 'react';
 import {useTap} from 'react-interactions/events/tap';
